rename `diff_btree_map` file to `delta_btree_map`
