+++ src/stream/src/executor/over_window/mod.rs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-mod diff_btree_map;
+mod delta_btree_map;
 mod eowc;
 mod general;
 mod state;
